Monstrous jankety commit; symbol table creation.
All instance declarations are now located in Types.hs, which makes
type stuff far (far) easier. The important stuff is the creation of the
symbol table, which is done by recursively walking down the AST while
maintaining the state of the table as things are added to it.

This might have been nice with a state monad; whatever.